# benchmark: TwoPC commit mode