--- a/go/test/endtoend/transaction/benchmark/schema.sql
+++ b/go/test/endtoend/transaction/benchmark/schema.sql
@@ -1,0 +1,5 @@
+create table test (
+    id bigint,
+    msg varchar(25),
+    primary key (id)
+) Engine=InnoDB;