<!--
  Thank you for your contribution to the Vitess project.
  How to contribute: https://vitess.io/docs/contributing/
  Please first make sure there is an open Issue to discuss the feature/fix suggested in this PR.
  If this is a new feature, please mark the Issue as "RFC".
 -->

<!-- if this PR is Work in Progress please create it as a Draft Pull Request -->

## Description

<!-- A few sentences describing the overall goals of the pull request's commits. -->
<!-- If this is a bug fix and you think the fix should be backported, please write so. -->

This PR benchmarks two Commit Modes:
1. Best Effort Commit (multi) – commits sequentially across shards.
2. Atomic Commit (TwoPC) – ensures atomicity through a multi-step process with parallel phases.

The benchmarks measure performance across transactions involving varying numbers of shards. 
Results are provided for 4, 8, and 16 shard configurations. 

4 shards
```
                          │ multi_v4.txt │             twopc_v4.txt             │
                          │    sec/op    │    sec/op     vs base                │
TwoPCCommit/SingleShard-8    756.5µ ± 1%   764.6µ ±  2%         ~ (p=0.180 n=6)
TwoPCCommit/TwoShards-8      1.545m ± 4%   3.386m ±  4%  +119.08% (p=0.002 n=6)
TwoPCCommit/ThreeShards-8    2.393m ± 2%   4.877m ± 14%  +103.82% (p=0.002 n=6)
TwoPCCommit/FourShards-8     3.309m ± 4%   6.098m ±  6%   +84.29% (p=0.002 n=6)
geomean                      1.744m        2.962m         +69.82%
```

8 shards
```
                          │ multi_8shards.txt │          twopc_8shards.txt           │
                          │      sec/op       │    sec/op     vs base                │
TwoPCCommit/SingleShard-8        750.3µ ±  1%    754.0µ ± 1%         ~ (p=0.818 n=6)
TwoPCCommit/TwoShards-8          1.528m ±  8%    3.448m ± 4%  +125.63% (p=0.002 n=6)
TwoPCCommit/FourShards-8         3.269m ±  3%    6.149m ± 5%   +88.10% (p=0.002 n=6)
TwoPCCommit/SixShards-8          5.176m ±  6%    8.577m ± 7%   +65.71% (p=0.002 n=6)
TwoPCCommit/EightShards-8        7.140m ± 11%   10.968m ± 5%   +53.62% (p=0.002 n=6)
geomean                          2.681m          4.320m        +61.12%
```



16 shards
```
                            │ multi_16shards.txt │          twopc_16shards.txt          │
                            │       sec/op       │    sec/op      vs base               │
TwoPCCommit/EightShards-8           7.611m ±  6%   11.261m ± 11%  +47.97% (p=0.002 n=6)
TwoPCCommit/TwelveShard-8           12.19m ± 10%    17.21m ± 11%  +41.18% (p=0.002 n=6)
TwoPCCommit/SixteenShards-8         17.35m ± 15%    24.06m ± 10%  +38.69% (p=0.002 n=6)
geomean                             11.72m          16.71m        +42.56%
```

Summary
* Best Effort Commit scales poorly as the number of shards increases due to its sequential execution of commits.
* Atomic Commit (TwoPC) incurs higher overhead but offers atomicity guarantees. Its parallel phases help reduce the performance gap as the shard count grows.


## Related Issue(s)

<!-- List related issues and pull requests. If this PR fixes an issue, please add it using Fixes #????  -->
- Closes #16245 

## Checklist

-   [ ] Tests were added or are not required
-   [ ] Did the new or modified tests pass consistently locally and on CI?
-   [ ] Documentation was added or is not required

## Deployment Notes

<!-- Notes regarding deployment of the contained body of work. These should note any db migrations, etc. -->